+++ sw/device/lib/runtime/print.h
@@ -8,11 +8,7 @@
 #include <stdarg.h>
 #include <stddef.h>
 
-#include "sw/device/lib/dif/dif_gpio.h"
-#include "sw/device/lib/dif/dif_spi_device.h"
 #include "sw/device/lib/dif/dif_uart.h"
-
-#include "spi_device_regs.h"  // Generated.
 
 /**
  * @file
@@ -50,30 +46,6 @@ typedef struct buffer_sink {
   void *data;
   sink_func_ptr sink;
 } buffer_sink_t;
-
-/**
- * SPI console buffer management constants.
- */
-enum {
-  kSpiDeviceReadBufferSizeBytes =
-      SPI_DEVICE_PARAM_SRAM_READ_BUFFER_DEPTH * sizeof(uint32_t),
-  kSpiDeviceFrameHeaderSizeBytes = 12,
-  kSpiDeviceBufferPreservedSizeBytes = kSpiDeviceFrameHeaderSizeBytes,
-  kSpiDeviceMaxFramePayloadSizeBytes = kSpiDeviceReadBufferSizeBytes -
-                                       kSpiDeviceFrameHeaderSizeBytes -
-                                       kSpiDeviceBufferPreservedSizeBytes - 4,
-  kSpiDeviceFrameMagicNumber = 0xa5a5beef,
-};
-
-/**
- * Returns a function pointer to the spi device sink function.
- */
-sink_func_ptr get_spi_device_sink(void);
-
-/**
- * Returns a function pointer to the uart sink function.
- */
-sink_func_ptr get_uart_sink(void);
 
 /**
  * Prints out a message to stdout, formatted according to the format string
@@ -314,32 +286,15 @@ size_t base_fhexdump_with(buffer_sink_t out, base_hexdump_fmt_t fmt,
 void base_set_stdout(buffer_sink_t out);
 
 /**
- * Configures SPI device GPIO TX indicator pin for `base_print.h` to use.
- *
- * Note that this function will save `gpio` in a global variable, so the
- * pointer must have static storage duration.
- *
- * @param gpio The GPIO handle to use for the SPI console TX indicator pin.
- * @param tx_indicator_pin The GPIO pin to use for the SPI console TX indicator.
- */
-void base_spi_device_set_gpio_tx_indicator(dif_gpio_t *gpio,
-                                           dif_gpio_pin_t tx_indicator_pin);
-
-/**
- * Configures SPI device stdout for `base_print.h` to use.
- *
- * Note that this function will save `spi_device` in a global variable, so the
- * pointer must have static storage duration.
- *
- * @param spi_device The SPI device handle to use for stdout.
- */
-void base_spi_device_stdout(const dif_spi_device_handle_t *spi_device);
-
-/**
- * Configures UART stdout for `base_print.h` to use.
+ * Configures a minimal stdout UART driver for `base_print.h` to use.
  *
  * Note that this function will save `uart` in a global variable, so the pointer
- * must have static storage duration.
+ * must have static storage duration. This driver uses
+ * `dif_uart_byte_send_polled` to send characters one by one, does not perform
+ * flow control and is generally unsafe to use in an interrupt setting.
+ *
+ * The UART IP and pinmux must have been initialized prior to calling this
+ * function.
  *
  * @param uart The UART handle to use for stdout.
  */
