+++ sw/device/lib/testing/test_framework/ottf_console.h
@@ -8,6 +8,7 @@
 #include <stdint.h>
 
 #include "sw/device/lib/base/status.h"
+#include "sw/device/lib/dif/dif_gpio.h"
 #include "sw/device/lib/dif/dif_uart.h"
 
 /**
